src/test/javascript/spec/app/entities/communication-configuration/communication-configuration.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { CommunicationConfigurationService } from 'app/entities/communication-configuration/communication-configuration.service';
import { ICommunicationConfiguration, CommunicationConfiguration } from 'app/shared/model/communication-configuration.model';
import { CommunicationMode } from 'app/shared/model/enumerations/communication-mode.model';

describe('Service Tests', () => {
  describe('CommunicationConfiguration Service', () => {
    let injector: TestBed;
    let service: CommunicationConfigurationService;
    let httpMock: HttpTestingController;
    let elemDefault: ICommunicationConfiguration;
    let expectedResult: ICommunicationConfiguration | ICommunicationConfiguration[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(CommunicationConfigurationService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new CommunicationConfiguration(0, 'AAAAAAA', 'AAAAAAA', CommunicationMode.MAIL, 'AAAAAAA', currentDate, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            createdAt: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a CommunicationConfiguration', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            createdAt: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.create(new CommunicationConfiguration()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a CommunicationConfiguration', () => {
        const returnedFromService = Object.assign(
          {
            priority: 'BBBBBB',
            retries: 'BBBBBB',
            mode: 'BBBBBB',
            name: 'BBBBBB',
            createdAt: currentDate.format(DATE_TIME_FORMAT),
            createdBy: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of CommunicationConfiguration', () => {
        const returnedFromService = Object.assign(
          {
            priority: 'BBBBBB',
            retries: 'BBBBBB',
            mode: 'BBBBBB',
            name: 'BBBBBB',
            createdAt: currentDate.format(DATE_TIME_FORMAT),
            createdBy: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a CommunicationConfiguration', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
